refactor(cli): extract self-contained termchart module
Move the ANSI chart renderers out of shared/ansiChart.ts into a
dedicated src/termchart/ module with no HyperDX imports (only
dependency: chalk), so the charting layer can be open-sourced later:

- types.ts       — TimeChartData / CategoricalEntry / TableColumn
- ansi.ts        — stripAnsi, colors, truncation, legend, time axis
- scale.ts       — resampleSeries / resampleLinear / niceTicks
- timeseries.ts  — renderLineChart / renderStackedBarChart
- charts.ts      — categorical / number / table / markdown renderers

Number formatting is now injected instead of imported: renderers take
formatTick / formatValue / formatNumericCell callbacks with a dep-free
compact default, and the HyperDX web-parity semantics (numbro formats,
axis ticks with average+mantissa:0) live in shared/formatNumber.ts and
shared/tileRender.ts closures. renderNumberChart takes pre-formatted
text. Data types now flow shared/* → termchart, never the reverse.

No behavior change — all 60 unit tests pass unchanged.
