+++ packages/cli/src/sourcemaps.ts
@@ -18,7 +18,7 @@ const log = (msg: string) => process.stdout.write(`${msg}\n`);
 const logError = (msg: string) => process.stderr.write(`${msg}\n`);
 
 const require = createRequire(import.meta.url);
-// eslint-disable-next-line no-restricted-imports -- package.json lives outside src, no @/ alias reaches it
+
 const PKG_VERSION: string = (require('../package.json') as { version: string })
   .version;
